requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
langchain-huggingface
flask
python-dotenv